Created new files in RPiHDMI folder for splitting functionality into their proper home and enabling raspi 4 support
