+++ workbench/devs/AHI/Drivers/RPiHDMI/Makefile.in
@@ -9,6 +9,6 @@ srcdir		= @srcdir@
 DRIVER		= rpihdmi.audio
 MODEFILE	= RPIHDMI
 
-OBJECTS		= rpihdmi-init.o rpihdmi-main.o rpihdmi-playslave.o rpihdmi-accel.o rpihdmi-hwaccess.o
+OBJECTS		= rpihdmi-init.o rpihdmi-main.o rpihdmi-playslave.o rpihdmi-accel.o rpihdmi-hwaccess.o rpihdmi-iec958.o
 
 include ../Common/Makefile.common
